Adds text search and range query profiles
Adds support for text search and range query profiles to the benchmark tool.

This introduces a `--query-profile` option to allow users to specify the type of query to run for workloads such as `stackoverflow-queries` and `query-users-by-name`.

Adds `text-prefix`, `text-search`, `text-search-rare`, `text-search-common`, and `text-search-mixed` options for StackOverflow text-based queries and `range` option for `query-users-by-name`.

Also, adds enhancements for accurate network utilization and telemetry for query benchmarks.